Refactor: ItemMainImg 분리
